pyproject.toml
[project]
name = "hikari-tsukino-bot"
version = "0.1.0"
description = "Hikari Tsukino — personal Telegram companion bot with memory"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"

dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
